// src/utils/types/global.ts
export type TProductSize = 'S' | 'M' | 'L' | 'XL' | '2XL' | '3XL'

export type TProductColor = {
  title: string
  value: string
}

export type TPrintAreaInfo = {
  // productImageId: string
  id: number
  area: {
    printX?: number // pixel position x
    printY?: number // pixel position y
    printW?: number // pixel width
    printH?: number // pixel height
    widthRealPx: number // real width of product image (for exporting)
    heightRealPx: number // real height of product image (for exporting)
  }
  surfaceType: TSurfaceType
  imageUrl: string
}

export type TProductMockupInfo = {
  variantId: number
  surfaceId: number
  mockupUrl: string
}

export type TBaseProduct = {
  id: number
  url: string
  name: string
  description: string
  variants: TClientProductVariant[] // variants
  detailImages: string[]
  inNewLine: boolean
  printAreaList: TPrintAreaInfo[] // surfaces
  mockups: TProductMockupInfo[]
}

export type TProductCategory =
  | 'shirt'
  | 'hat'
  | 'cup'
  | 'keychain'
  | 'phonecase'
  | 'figurine'
  | 'totebag'

export type TClientProductVariant = {
  id: number
  name: string
  size: TProductSize
  color: TProductColor
  priceAmountOneSide: number
  priceAmountBothSide: number
  currency: string
  priceAfterDiscount?: number
  stock: number
  category?: TProductCategory
}
// ...
export type TSurfaceType = 'front' | 'back' | 'both'
// ...
export type TFramesCount = 1 | 2 | 3 | 4

export type TTemplateType =
  | `1-horizon`
  | '1-vertical'
  | '1-square'
  | '2-horizon'
  | '2-vertical'
  | '3-left'
  | '3-right'
  | '3-top'
  | '3-bottom'
  | '4-square'
  | '4-horizon'

export type TPrintTemplate = {
  id: string
  name: string
  frames: TTemplateFrame[]
  type: TTemplateType
  framesCount: TFramesCount
}

export type TFrameRectType = 'horizontal' | 'vertical' | 'square'

export type TTemplateFrame = {
  id: string
  index: number
  placedImage?: TPlacedImage
  rectType: TFrameRectType
}

export type TPlacedImage = {
  id: string
  placementState: TPlacementState
  imgURL: string
}